+++ packages/explorerkit-translator/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @solanafm/explorer-kit
 
+## 1.1.3
+
+### Patch Changes
+
+- b1ca99f: feat: add support for optional account keys
+
+## 1.1.2
+
+### Patch Changes
+
+- d833b78: feat: add support for optional account keys
+- daa0990: fix: patch token extensions account key mapping
+
+## 1.1.1
+
+### Patch Changes
+
+- 30f749f: support shank event parsing
+
+## 1.1.0
+
+### Minor Changes
+
+- f666e44: feat: add support for anchorV1 idls (0.3.0)
+
 ## 1.0.5
 
 ### Patch Changes